move reusable code to utils
